Everything I thought I knew about sides was a lie
Apparently @SideOnly is bad, and having an isRemote check isn't enough sometimes. Will document this stuff later on.
