+++ src/main/java/me/fril/regeneration/util/RenderUtil.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 import org.lwjgl.opengl.GL11;
 
-import me.fril.regeneration.client.models.ModelArmorOverride;
+import me.fril.regeneration.client.rendering.ModelArmorOverride;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.entity.AbstractClientPlayer;
 import net.minecraft.client.model.ModelPlayer;
